backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.4
Flask-SocketIO==5.1.1
python-socketio==5.4.0
redis==4.0.2
SQLAlchemy==1.4.23
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.3.0
requests==2.26.0
gunicorn==20.1.0
alembic==1.7.4 